Fix `github_branch_protection` migration test to be a unit test
